Fix bug when requested element id is null in ElementRelationsController::actionGetByElementId. `elementId` in  is now always an int. (0 = null). #21

+++ src/controllers/ElementRelationsController.php
@@ -17,7 +17,7 @@ class ElementRelationsController extends Controller
 
     public function actionGetByElementId(): string
     {
-        $elementId = Craft::$app->request->getParam('elementId');
+        $elementId = (int) Craft::$app->request->getParam('elementId');
         $siteId = Craft::$app->request->getParam('siteId');
         $force = Craft::$app->request->getParam('force') === 'true';
         $elementRelations = CacheService::getElementRelationsCached($elementId, $force);
